requirements.txt
streamlit==1.54.0
pandas==2.3.3
pillow==12.1.1
loguru==0.7.3
python-dotenv==1.2.1
requests==2.32.5
altair==6.0.0
